Changes key bindings to be more like magit & dired
l in dired redisplays at point; opening the pivotal projects buffer is
more like ^, which navigates to the parent directory in dired.

t in dired toggles all marks between on/off state. C-m, in both dired
and magit, opens something in a new buffer. pivotal-toggle-visibility is
more like dired's i, which inserts a subdirectory, or magit's tab, which
reveals more details about whatever is a point.